refactor: split up process instance mocks
