+++ operate/client/e2e-playwright/docs-screenshots/get-familiar-with-operate.spec.ts
@@ -25,7 +25,7 @@ import {
 import {
   mockResponses as mockProcessInstanceDetailResponses,
   runningOrderProcessInstance,
-} from '../mocks/processInstance.mocks';
+} from '../mocks/processInstance';
 import {open} from 'modules/mocks/diagrams';
 
 test.beforeEach(async ({page, commonPage, context}) => {
