+++ operate/client/e2e-playwright/docs-screenshots/variables-and-incidents.spec.ts
@@ -13,7 +13,7 @@ import {mockResponses as mockProcessesResponses} from '../mocks/processes.mocks'
 import {
   mockResponses as mockProcessDetailResponses,
   orderProcessInstance,
-} from '../mocks/processInstance.mocks';
+} from '../mocks/processInstance';
 
 test.beforeEach(async ({page, commonPage, context}) => {
   await commonPage.mockClientConfig(context);
